Initial commit of plugin stuff. The basic-est of flows is working.
Creating a new note works. Writing it to a directory. Opening it.
Assigning it some settings. A "is Markdown" syntax file. The basic-est
of settings. The bones are mostly here.

I think the next step is to create a TextInputHandler/ListInputHandler
to do a Notational-Velocity-style search through existing notes; that
means the title will be defaulted to exist.